project_2/workspace/requirements.txt
absl-py==0.8.1
asn1crypto==1.2.0
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
cachetools==4.2.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
dill==0.3.1.1
flatbuffers==2.0
future==0.18.2
gast==0.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.41.1
h5py==2.9.0
idna==2.8
importlib-resources==5.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
jupyter-client==5.3.4
jupyter-core==4.6.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
libclang==12.0.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.1.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
promise==2.2.1
prompt-toolkit==3.0.2
protobuf==3.19.1
psutil==5.6.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
pyspark==2.4.3
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.3.2
sip==4.19.13
six==1.13.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.7.0
tensorflow-io-gcs-filesystem==0.21.0
tensorflow-metadata==0.14.0
termcolor==1.1.0
tfds-nightly==4.4.0.dev202111090109
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.6.0
